Reorganize git_config module
